requirements.txt
fastapi
uvicorn 
python-dotenv
pyjwt
asgiref==3.4.1
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
h11==0.12.0
pycparser==2.20
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.0
pymongo
motor
pymongo[srv] 
